requirements.txt
Babel==2.3.4
blinker==1.3
click==6.6
decorator==4.0.10
Flask==0.10.1
Flask-Babel==0.11.1
Flask-Login==0.4.0
Flask-SQLAlchemy==1.0
Flask-WhooshAlchemy==0.8
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.9
pbr==1.10.0
pytz==2016.7
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.3
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.11
Whoosh==2.6.0
WTForms==2.1
